+++ src/ensembl/stories/shared-components/image-button/ImageButton.stories.scss
