+++ src/ensembl/stories/shared-components/index.ts
@@ -2,4 +2,4 @@ import './close-button/CloseButton.stories';
 import './question-button/QuestionButton.stories';
 import './dropdown/Dropdown.stories';
 import './input/Input.stories';
-import './toggle-image-button/ToggleImageButton.stories';
+import './image-button/ImageButton.stories';
